# swagger/swagger.json
{"swagger":"2.0","info":{"title":"Data Table API","description":"Simple CRUD API implemented with goa","version":""},"host":"localhost:5000","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/api/users":{"get":{"tags":["user"],"summary":"list user","description":"Retrieve all users.","operationId":"user#list","produces":["application/data.table.user; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataTableUserCollection"}}},"schemes":["http"]},"post":{"tags":["user"],"summary":"create user","description":"Create new user","operationId":"user#create","produces":["application/vnd.goa.error"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/CreateUserPayload"}}],"responses":{"201":{"description":"Created"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/api/users/{userID}":{"get":{"tags":["user"],"summary":"show user","description":"Retrieve user with given id.","operationId":"user#show","produces":["application/data.table.user"],"parameters":[{"name":"userID","in":"path","description":"User ID","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/DataTableUser"}},"404":{"description":"Not Found"}},"schemes":["http"]},"put":{"tags":["user"],"summary":"update user","description":"Update a user","operationId":"user#update","produces":["application/vnd.goa.error"],"parameters":[{"name":"userID","in":"path","description":"User ID","required":true,"type":"string"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/UpdateUserPayload"}}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]},"delete":{"tags":["user"],"summary":"delete user","operationId":"user#delete","produces":["application/vnd.goa.error"],"parameters":[{"name":"userID","in":"path","description":"User ID","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download public/swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}}},"definitions":{"CreateUserPayload":{"title":"CreateUserPayload","type":"object","properties":{"age":{"type":"integer","description":"Age","example":25,"minimum":18,"maximum":150},"firstName":{"type":"string","description":"First Name","example":"John"},"lastName":{"type":"string","description":"Last Name","example":"Doe"}},"example":{"age":25,"firstName":"John","lastName":"Doe"},"required":["firstName","lastName"]},"DataTableUser":{"title":"Mediatype identifier: application/data.table.user; view=default","type":"object","properties":{"age":{"type":"integer","description":"Age","example":25,"minimum":18,"maximum":150},"firstName":{"type":"string","description":"First Name","example":"John"},"id":{"type":"string","description":"User unique identifier","example":"123e4567-e89b-12d3-a456-426655440000"},"lastName":{"type":"string","description":"Last Name","example":"Doe"}},"description":"A user (default view)","example":{"age":25,"firstName":"John","id":"123e4567-e89b-12d3-a456-426655440000","lastName":"Doe"},"required":["id","firstName","lastName"]},"DataTableUserCollection":{"title":"Mediatype identifier: application/data.table.user; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/DataTableUser"},"description":"DataTableUserCollection is the media type for an array of DataTableUser (default view)","example":[{"age":25,"firstName":"John","id":"123e4567-e89b-12d3-a456-426655440000","lastName":"Doe"},{"age":25,"firstName":"John","id":"123e4567-e89b-12d3-a456-426655440000","lastName":"Doe"}]},"UpdateUserPayload":{"title":"UpdateUserPayload","type":"object","properties":{"age":{"type":"integer","description":"Age","example":25,"minimum":18,"maximum":150},"firstName":{"type":"string","description":"First Name","example":"John"},"lastName":{"type":"string","description":"Last Name","example":"Doe"}},"example":{"age":25,"firstName":"John","lastName":"Doe"}},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"Created":{"description":"Created"},"NoContent":{"description":"No Content"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/DataTableUser"}}}}